Modified ConversationLoop to take a reader and writer.
Instead of taking input directly from os.Stdin and sending output
directly to os.Stdout, ConversationLoop now takes an io.Reader and
io.Writer type and lets the caller decide what the reader and
writer is to be. This facilitates testing the ConversationLoop
and also should expand options down the road.