Fix state transition navigation.
State transitions navigation was broken. Fixed it by moving state
transition to the begining of the action sequence. That involved
many changes to the way states are organized, but the transition
sequence is much more straightforward now.

Also added some debugging tools to help understand and troubleshoot
navigation through the state machine.